+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 0.9.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Bugfix: use DEM epsg_code for 2D models.
 
 
 0.9.0 (2021-12-15)