+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog of threedigrid-builder
 
 - Bugfix: use DEM epsg_code for 2D models.
 
+- Bugfix: Small fix for lgrtot
+
 
 0.9.0 (2021-12-15)
 ------------------